[IRIS-1330] Don't break conversations db on RENAME
At the tail end of the RENAME command, mailbox_rename_cleanup() was
calling mailbox_delete() on the old copy of the folder which had just
been renamed.  Mailbox_delete() updated the conversations db as if the
folder had been deleted, bumping down exists counts etc.  Unfortunately,
the exists counts had *not* been bumped up when the folder was copied
earlier in the rename process; instead conversation_rename_folder() had
rewritten the B records with the same counts but the new folder name.

The effect was to at least break the counts on every B record touched by
the folder.  Conversations existing only in that one folder had their B
record deleted.

The quick and ugly solution is to pass an extra parameter to
mailbox_delete().  Better solutions are invited!