[IRIS-1399] fix conversation_find_folder
to return either the correct folder or NULL.

Also, don't add unnecessary entries to $FOLDER_NAMES if
conversation_find_folder() is called with a folder name which does not
exist.  This mboxname comes in directly off the wire and could be any
old rubbish, not necessarily a current or future valid folder name.