[IRIS-1466] match NULL/empty subject correctly
A NULL or empty Subject header field in a message does not match any
non-empty subject in an S record, it only matches empty subjects or
subjects which normalise to empty such as "Re:".

We still handle a missing S-record as a special match-anything case
which indicates that the conversations DB has not been reconstructed
since the IRIS1466 code was deployed, and which therefore matches any
incoming subject.