Don't stop early when auditing corrupted DBs
If a conversations DB has a single corrupted record, do not stop the
audit early but continue to detect problems with other records.  We only
stop early if we encounter an error writing to the temporary DB, and
then we clearly say "giving up".

Also, print useful error messages when encountering failures in
fix_modseqs(), including the DB pathname and the record key.